.autoupdate/create-ps-dir/requirements.txt
requests == 2.26.0
PyGithub == 1.55